requirements.txt
asgiref==3.5.2
Django==3.2
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
importlib-metadata==4.12.0
Markdown==3.3.7
Pillow==9.2.0
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
zipp==3.8.1
